# Execute one query to retrieve all feature flag data